+++ lib/src/xendit/models/entity/xendit_actions.dart
@@ -24,6 +24,18 @@ class XenditActions {
   /// The generated QR code to be scanned by the end user in order to perform the action
   final String? qrCode;
 
+  /// desktop_web_checkout_url
+  final String? desktopWebCheckoutUrl;
+
+  /// mobile_web_checkout_url
+  final String? mobileWebCheckoutUrl;
+
+  /// mobile_deeplink_checkout_url
+  final String? mobileDeeplinkCheckoutUrl;
+
+  /// qr_checkout_string
+  final String? qrCheckoutString;
+
   /// If status=REQUIRES_ACTION, this contains objects that detail the possible next steps in order to complete a payment. Only one of the provided actions is required to be fulfilled. If no further action is needed, this parameter will be an empty array [].
   XenditActions({
     this.method,
@@ -31,6 +43,10 @@ class XenditActions {
     this.action,
     this.url,
     this.qrCode,
+    this.desktopWebCheckoutUrl,
+    this.mobileWebCheckoutUrl,
+    this.mobileDeeplinkCheckoutUrl,
+    this.qrCheckoutString,
   });
 
   factory XenditActions.fromJson(Map<String, dynamic> json) => XenditActions(
@@ -39,6 +55,10 @@ class XenditActions {
         action: json["action"],
         url: json["url"],
         qrCode: json["qr_code"],
+        desktopWebCheckoutUrl: json["desktop_web_checkout_url"],
+        mobileWebCheckoutUrl: json["mobile_web_checkout_url"],
+        mobileDeeplinkCheckoutUrl: json["mobile_deeplink_checkout_url"],
+        qrCheckoutString: json["qr_checkout_string"],
       );
 
   Map<String, dynamic> toJson() => {
@@ -47,5 +67,12 @@ class XenditActions {
         if (action != null) "action": action,
         if (url != null) "url": url,
         if (qrCode != null) "qr_code": qrCode,
+        if (desktopWebCheckoutUrl != null)
+          "desktop_web_checkout_url": desktopWebCheckoutUrl,
+        if (mobileWebCheckoutUrl != null)
+          "mobile_web_checkout_url": mobileWebCheckoutUrl,
+        if (mobileDeeplinkCheckoutUrl != null)
+          "mobile_deeplink_checkout_url": mobileDeeplinkCheckoutUrl,
+        if (qrCheckoutString != null) "qr_checkout_string": qrCheckoutString,
       };
 }