+++ CHANGELOG.md
@@ -19,5 +19,6 @@
 * Add debug selection for showing log.
 * Adding some documentation on the code.
 * Adding `success_redirect_url`, `failure_redirect_url`, and `cancel_redirect_url` on the Channel Properties.
+* Adding `desktop_web_checkout_url`, `mobile_web_checkout_url`, `mobile_deeplink_checkout_url`, `qr_checkout_string` on the Actions
 * Adding some `enums`.
 * Fix minor bugs.